rename: 버튼 컴포넌트 위치 이동 (#34)
