+++ src/shared/ui/index.ts
@@ -2,7 +2,7 @@ import { Toaster } from '@/shared/ui/Toaster'
 import Footer from '@/shared/ui/footer/Footer'
 import { Modal, ModalClose } from '@/shared/ui/Modal'
 import Header from '@/shared/ui/header/Header'
-import { Button, ButtonVariants } from '@/shared/ui/button/Button'
+import { Button, ButtonVariants } from '@/shared/ui/Button'
 import { Input } from '@/shared/ui/input/Input'
 
 export {
